Update README.md
Write a multithreaded program in C that outputs prime numbers. This program should
work as follows: The user will run the program and will enter a number on the command
line. The program will then create a separate thread that outputs all the prime numbers
less than or equal to the number entered by the user.